myshop/requirements.txt
braintree==4.12.0
celery==5.2.0
cryptography==36.0.0
Django==3.2.8
django-allauth==0.47.0
django-celery-beat==2.2.1
djangorestframework==3.13.1
flower==1.0.0
oauthlib==3.1.1
Pillow==8.4.0
psycopg2-binary==2.9.3
python-crontab==2.6.0
redis==4.0.2
requests==2.26.0
weasyprint==56.1
social-auth-app-django==5.0.0